All tests of "near equality" happen in one place
We want to ensure that `husl-numpy` is gives the same color conversions as the original `husl.py` project. Because we're working with floats of different precision (Python's native `float` and `np.float64`, for example), we won't have *exact* equality for every calculation. The `_diff` utilty function in the test suite ensures that we always compare floats to the same precision.