added script to generate a L,H->Chroma table
Calculating chroma for the saturation channel is by FAR the most expensive
step in RGB->HUSL. It's a good candidate for a lookup table. Even if we generate
a huge, cache-polluting table, it'll be faster than calculating arctangent, sine, cosine,
a few powers of odd floating point numbers, and so on, a bunch of fmin calls, etc.